Disable nulling of song info
After some thought I realized that resetting the song information leads to a lot of unwanted behaviour, and since we have a `song_state` flag to tell us whether a gameplay session is happening anyway, this is sufficient to tell us whether to display information or not.

I have also simplified the JS of the web app's JS. Now that we're no longer nulling song info, we don't need to concern ourselves with hiding or showing certain elements with a delay. We can instead set a single delay for the song state, and that will make the elements appear and disappear in a non-jarring way.

Also added quick update instructions to README.