Maintenance update
- Update dependencies
- Ensure Main Stage continues to work.

Known Issues:
- Battle Stage functionality is broken
- Certain conditions are no longer being picked up (for example matchmaking status).

The project has been unmaintained for a while, and since there have been several breaking changes to the log behavior, this needs to be overlooked carefully. For now, this will restore functionality to a working state in Main Stage only, but still has a ways to go until we're back to where this once was.